requirements.txt
appdirs==1.4.2
click==6.7
coreapi==2.2.4
coreapi-cli==1.0.6
coreschema==0.0.4
Django==1.10.5
djangorestframework==3.5.4
djangorestframework-jwt==1.9.0
httpie==0.9.9
itypes==1.1.0
Jinja2==2.9.5
MarkupSafe==0.23
packaging==16.8
Pygments==2.2.0
PyJWT==1.4.2
pyparsing==2.1.10
requests==2.13.0
six==1.10.0
uritemplate==3.0.0
